Update How Options are Registered with CLImber
Modifications to the CommandOptionAttribute class to allow options to be registered with an abbreviation instead of relying on multiple instances of the attribute on a property.
